requirements.txt
celery==5.2.3
Django==3.1.6
django-celery-beat==2.4.0
django-celery-results==2.4.0
gevent==22.10.2
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-environ==0.4.5
djangorestframework-simplejwt==5.2.2
mysql-connector-python==8.0.29
mysqlclient==2.1.1
selenium==4.7.2
webdriver-manager==3.8.5
openpyxl==3.0.10
redis==4.1.3
django-filter==21.1
pyotp==2.8.0
openai==0.19.0
Pillow==9.4.0
facebook-sdk==3.1.0
nltk==3.8.1
pycountry==22.3.5
xlsxwriter==3.1.0
reportlab==4.0.0
matplotlib==3.7.1
xlwt==1.3.0
pyxlsb==1.0.10
emoji==2.2.0
stripe==2.74.0
paypalrestsdk==1.13.1
google-search-results